noapp.kill added as one-shot app boot prevention (#41)
* noapp.kill file in app dir prevents boot of app

Added logic to check for the presence of /app/noapp.kill during boot. If the file exists, the script logs a message, deletes the file, and skips starting /app/app, similar to the existing gadget mode check.  This includes a self- rm -f to not brick if users somehow lose ssh access, use this in unintended ways, etc.

* Make rc.local +x

* Update rc.local

redundant logging removed

* Add newline to end of file

---------

Co-authored-by: Sims <38142618+suchmememanyskill@users.noreply.github.com>